update アクションごとに呼び出していたjsファイルの処理を全てdefault.jsに移行。
アクションごとのjsファイルは削除済
