--- app/frontend/entry/searches/users/index.js
@@ -1,10 +1,0 @@
-import FollowButton from "../../../js/components/follow_button";
-
-document.addEventListener("DOMContentLoaded", function () {
-  new vue({
-    el: "#container",
-    components: {
-      "follow-button": FollowButton
-    }
-  });
-});
