--- app/frontend/entry/settings/profile/index.js
@@ -1,10 +1,0 @@
-import ProfilePhotoForm from '../../../js/components/profile_photo_form';
-
-document.addEventListener('DOMContentLoaded', function () {
-  // new vue({
-  //   el: "#photo-form",
-  //   components: {
-  //     "profile-photo-form": ProfilePhotoForm
-  //   }
-  // })
-});
