+++ app/frontend/js/default.js
@@ -1,7 +1,37 @@
 require("selectize");
 import tippy from "tippy.js";
+import CircleFavoriteButton from "./components/circle_favorite_button";
+import DeleteCommentButton from "./components/delete_comment_button";
+import DeleteDraftButton from "./components/delete_draft_button";
+import FollowButton from "./components/follow_button";
+import FollowTagButton from "./components/follow_tag_button";
+import RememberButton from "./components/remember_button";
+
+const toast = function (type, message) {
+  Swal.fire({
+    type: type,
+    title: message,
+    toast: true,
+    position: 'bottom-start',
+    grow: 'column',
+    showConfirmButton: false,
+    showCloseButton: true
+  });
+}
 
 document.addEventListener('DOMContentLoaded', function () {
+  new vue({
+    el: "#container",
+    components: {
+      CircleFavoriteButton,
+      DeleteCommentButton,
+      DeleteDraftButton,
+      FollowButton,
+      FollowTagButton,
+      RememberButton
+    }
+  });
+
   // ドロップダウンメニュー
   const dropdown = document.getElementsByClassName("js-dropdown");
   const dropdownMenu = document.getElementsByClassName("js-dropdown-target");
@@ -19,7 +49,6 @@ document.addEventListener('DOMContentLoaded', function () {
       if (dropdown[index].contains(event.srcElement)) {
         continue;
       }
-
       const element = dropdownMenu[index];
       if (!element.className.includes("u-hide")) {
         element.classList.add("u-hide");
@@ -31,14 +60,7 @@ document.addEventListener('DOMContentLoaded', function () {
   const type = document.getElementById("toast_type");
   const message = document.getElementById("toast_message");
   if (type && message) {
-    Swal.fire({
-      type: type.value,
-      title: message.value,
-      toast: true,
-      position: 'bottom-start',
-      showConfirmButton: false,
-      showCloseButton: true
-    });
+    toast(type.value, message.value);
   }
 
   // ツールチップ
@@ -62,5 +84,11 @@ document.addEventListener('DOMContentLoaded', function () {
         text: input
       }
     }
+  });
+
+  //ujsのメッセージを表示
+  document.body.addEventListener("ajax:complete", function (event) {
+    var data = JSON.parse(event.detail[0].response);
+    toast(data.type, data.message);
   });
 });
